+++ src/Physics/PhysX/Joint.cpp
@@ -48,7 +48,7 @@ namespace hws::physx {
     ctx_.physx_mutex_.lock();
     assert(type_ != urdf::JointType_e::joint_fixed && "[Joint] fixed joint cannot be moved");
     if(px_axis_ != ::physx::PxArticulationAxis::eCOUNT)
-      px_base_->setJointPosition(px_axis_, std::fmod(direction_ * position, M_PI * 2));
+      px_base_->setJointPosition(px_axis_, (float)std::fmod(direction_ * position, M_PI * 2.));
     else
       assert(false && "[Joint floating and planar joints are not yet supported]");
     ctx_.physx_mutex_.unlock();
@@ -59,7 +59,7 @@ namespace hws::physx {
     ctx_.physx_mutex_.lock();
     assert(type_ != urdf::JointType_e::joint_fixed && "[Joint] fixed joint cannot be moved");
     if(px_axis_ != ::physx::PxArticulationAxis::eCOUNT)
-      px_base_->setJointVelocity(px_axis_, velocity);
+      px_base_->setJointVelocity(px_axis_, (float)velocity);
     else
       assert(false && "[Joint floating and planar joints are not yet supported]");
     ctx_.physx_mutex_.unlock();
